+++ src/posthog-core.ts
@@ -72,7 +72,7 @@ import { extendURLParams, request, SUPPORTS_REQUEST } from './request'
 import { Heatmaps } from './extensions/heatmaps'
 import { ScrollManager } from './scroll-manager'
 import { SimpleEventEmitter } from './utils/simple-event-emitter'
-import { Autocapture } from './extensions/autocapture'
+import { Autocapture } from './extensions/dom-autocapture/autocapture'
 import { TracingHeaders } from './extensions/tracing-headers'
 import { ConsentManager } from './consent'
 import { ExceptionObserver } from './extensions/exception-autocapture'
